requirements.txt
streamlit
sentence-transformers
faiss-cpu
huggingface_hub
fuzzywuzzy
pandas
numpy
python-Levenshtein  # Required for fuzzywuzzy
jupyter